wip: core/util stuff is kind of a mess.. moving things around (the error that appeared is a socket.gethostbyname call -- there's probably something deeper, but appending '.local' to the hostname seems to fix things.. it shouldn't break to begin with, so isolating some things)
